requirements/base.txt
begins
mido
python-rtmidi
